Moved Signature components into plugin dir.
