+++ plugins/signatureCanvas/SignatureCanvas.tsx
